requirements/base.txt
Django==1.11.5
djangorestframework==3.6.4
pytz==2017.2
djangorestframework-recursive==0.1.2
coreapi-cli==1.0.6
